src/server/main/requirements.txt
fastapi
uvicorn
python-dotenv
motor
httpx
mem0ai
qwen-agent

